many minor changes addressing strict compiler warnings
git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aro/trunk@1174 7b6de600-d9ab-45c9-a1ca-e8145d3c6daf

+++ ivp/src/lib_bhvutil/WaypointEngine.cpp
@@ -233,7 +233,7 @@ string WaypointEngine::setNextWaypoint(double os_x, double os_y)
   // (3) declare completion of the behavior.
   if(point_advance) {
     m_curr_ix++;
-    if(m_curr_ix >= vsize) {
+    if(m_curr_ix >= (int)(vsize)) {
       m_curr_ix = 0;
       if(!m_perpetual) {
 	if(m_repeat > 0)
